sample updates
