+++ src/main/java/com/truelayer/java/TrueLayerClient.java
@@ -2,9 +2,9 @@ package com.truelayer.java;
 
 import com.truelayer.java.auth.IAuthenticationHandler;
 import com.truelayer.java.hpp.IHostedPaymentPageLinkBuilder;
+import com.truelayer.java.mandates.IMandatesApi;
 import com.truelayer.java.merchantaccounts.IMerchantAccountsApi;
 import com.truelayer.java.payments.IPaymentsApi;
-import com.truelayer.java.recurringpayments.IMandatesApi;
 import lombok.AllArgsConstructor;
 import org.apache.commons.lang3.ObjectUtils;
