+++ src/main/java/com/truelayer/java/mandates/entities/mandatedetail/AuthorizedMandateDetail.java
@@ -1,4 +1,4 @@
-package com.truelayer.java.recurringpayments.entities.mandatedetail;
+package com.truelayer.java.mandates.entities.mandatedetail;
 
 import com.truelayer.java.entities.AuthorizationFlowWithConfiguration;
 import lombok.EqualsAndHashCode;
