+++ src/main/java/com/truelayer/java/mandates/entities/mandatedetail/AuthorizingMandateDetail.java
@@ -1,4 +1,4 @@
-package com.truelayer.java.recurringpayments.entities.mandatedetail;
+package com.truelayer.java.mandates.entities.mandatedetail;
 
 import com.truelayer.java.entities.AuthorizationFlowWithConfiguration;
 import lombok.EqualsAndHashCode;
@@ -6,11 +6,9 @@ import lombok.Value;
 
 @Value
 @EqualsAndHashCode(callSuper = false)
-public class RevokedMandateDetail extends MandateDetail {
+public class AuthorizingMandateDetail extends MandateDetail {
 
-    Status status = Status.REVOKED;
-
-    String revocationSource;
+    Status status = Status.AUTHORIZING;
 
     AuthorizationFlowWithConfiguration authorizationFlow;
 }
