+++ src/main/java/com/truelayer/java/mandates/entities/mandatedetail/RevokedMandateDetail.java
@@ -1,4 +1,4 @@
-package com.truelayer.java.recurringpayments.entities.mandatedetail;
+package com.truelayer.java.mandates.entities.mandatedetail;
 
 import com.truelayer.java.entities.AuthorizationFlowWithConfiguration;
 import lombok.EqualsAndHashCode;
@@ -6,9 +6,11 @@ import lombok.Value;
 
 @Value
 @EqualsAndHashCode(callSuper = false)
-public class AuthorizingMandateDetail extends MandateDetail {
+public class RevokedMandateDetail extends MandateDetail {
 
-    Status status = Status.AUTHORIZING;
+    Status status = Status.REVOKED;
+
+    String revocationSource;
 
     AuthorizationFlowWithConfiguration authorizationFlow;
 }
