+++ src/test/java/com/truelayer/java/integration/MandatesIntegrationTests.java
@@ -7,9 +7,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import com.truelayer.java.TestUtils;
 import com.truelayer.java.TestUtils.RequestStub;
 import com.truelayer.java.http.entities.ApiResponse;
-import com.truelayer.java.payments.entities.*;
-import com.truelayer.java.recurringpayments.entities.CreateMandateRequest;
-import com.truelayer.java.recurringpayments.entities.CreateMandateResponse;
+import com.truelayer.java.mandates.entities.CreateMandateRequest;
+import com.truelayer.java.mandates.entities.CreateMandateResponse;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
